split logic linked mandatarissen in controller and data-access
